connected deleted route, and added functionality to delete a seminar
